requirements.txt
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
fastapi==0.65.1
google-api-core==1.28.0
google-auth==1.30.1
google-cloud-core==1.6.0
google-cloud-datastore==1.15.3
google-cloud-ndb==1.8.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
h11==0.12.0
httptools==0.1.2
idna==2.10
packaging==20.9
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pymemcache==3.4.3
pyparsing==2.4.7
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
